cstr_kafka_influxdb/model/requirements.txt
faust-streaming
numpy
scipy
aiokafka>=0.9.0,<0.11.0